Simple train script to be used on the car
Requires a command line argument, which is the path you want to write jpeg images to. Other than that, it is just a while loop that has you press enter to take an image, which is then written to the relevant folder path. Keeps going until you end the script via ctrl+c